requirements.txt
pandas==2.2.2
numpy==1.26.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
scikit-learn==1.3.1
matplotlib==3.8.1
seaborn==0.12.2
Faker==23.11.1
